feat: Implement task selection modal with drag-and-drop functionality and dynamic task management
